First crack at the multiplayer component registry to allow multiplayer components to live in any gem

+++ Gems/Multiplayer/Code/Source/NetworkEntity/NetworkEntityRpcMessage.cpp
@@ -42,7 +42,7 @@ namespace Multiplayer
         }
     }
 
-    NetworkEntityRpcMessage::NetworkEntityRpcMessage(RpcDeliveryType rpcDeliveryType, NetEntityId entityId, NetComponentId componentId, uint8_t rpcMessageType, ReliabilityType isReliable)
+    NetworkEntityRpcMessage::NetworkEntityRpcMessage(RpcDeliveryType rpcDeliveryType, NetEntityId entityId, NetComponentId componentId, uint16_t rpcMessageType, ReliabilityType isReliable)
         : m_rpcDeliveryType(rpcDeliveryType)
         , m_entityId(entityId)
         , m_componentId(componentId)
@@ -98,7 +98,7 @@ namespace Multiplayer
         static constexpr uint32_t sizeOfFields = sizeof(RpcDeliveryType)
             + sizeof(NetEntityId)
             + sizeof(NetComponentId)
-            + sizeof(uint8_t);
+            + sizeof(uint16_t);
 
         // 2-byte size header + the actual blob payload itself
         const uint32_t sizeOfBlob = (m_data != nullptr) ? sizeof(uint16_t) + m_data->GetSize() : 0;
@@ -127,7 +127,7 @@ namespace Multiplayer
         return m_componentId;
     }
 
-    uint8_t NetworkEntityRpcMessage::GetRpcMessageType() const
+    uint16_t NetworkEntityRpcMessage::GetRpcMessageType() const
     {
         return m_rpcMessageType;
     }
